Add Neovim config to the repository
This is a minimal, single-file Neovim config written in Lua, which is
heavily inpsired by kickstart.nvim. I've tried to keep everything very
simple, sensible, and as close to the default Vim experience as
possible. Essentially, this is just vanilla Neovim with a handful of
what I consider to be the most important plugins: lspconfig,
treesitter, telescope, and a few other things.

Hopefully, this config succeeds in extending the built-in capabilities
of Neovim without falling into the trap of trying to turn it into
another VS Code clone.