Add Alacritty config to the repository
Although I've had no complaints with either iTerm2 or GNOME Terminal,
I think it would be nice to use the same terminal on both Mac and Linux.
The two most popular options for a modern cross-platform terminal seem
to be Alacritty and Kitty, and after playing around with both, I think
Alacritty is more my style.